ASABASE-343 - segmentation fault when trying to run Timer
1) Ensured that values with signature 'b' (aka booleans) are
   unmarshalled into 4 bytes long variables.

2) Removed unnecessary calls to AJ_CloseMsg(). There is a catch-all
   call in the application message loop.

+++ time/src/TimeServiceServer/TimeServiceServer.c
@@ -661,7 +661,7 @@ AJ_Status AJTS_Server_PropSetHandler(AJ_Message* replyMsg, uint32_t propId, void
 
     case TIME_ALARM1_ENABLED_PROP:
         {
-            uint8_t isEnabled;
+            uint32_t isEnabled;
 
             status = AJ_UnmarshalArgs(replyMsg, "b", &isEnabled);
             if (status != AJ_OK) {
